Introduced Lobby
Introduced the lobby, modified the chat javascript in order to use our
communicator, modified the build system by removing the useless result
folder, updated the image location to point at a fixed dropbox link,
updated the images.
Up to now, the lobby works with a general chat that can be used.