Second iteration of the lobby
The rooms are correctly destroyed once the game has ended, the list is
automatically updated and removed based on the number of the players
playing. Still problems with the naming of the rooms, will be fixed in
the next iteration. Fixed also a problem with the chat conflicting with
the existing chat modules.